YP-21 feat: добавлено управление
добавлена update функция позволяющая отслеживать мультинажатия на перемещение, теперь можно двигаться наискось
