+++ packages/client/src/Components/App/App.tsx
@@ -5,13 +5,10 @@ import { Outlet } from 'react-router-dom';
 import { DebugPanel, Menu } from '@Components';
 import { useRender, useUpdate } from '@Game';
 import { useGameLoop, useGetCanvasSize, useSetCanvasContext } from '@Hooks';
-import { movePlayer, useAppDispatch } from '@Store';
 
 import './App.scss';
 
 export const App = () => {
-  const dispatch = useAppDispatch();
-
   const canvasRef = useRef<HTMLCanvasElement>(null);
   const [width, height] = useGetCanvasSize();
   const ctx = useSetCanvasContext(canvasRef);
@@ -34,33 +31,6 @@ export const App = () => {
       stopGame();
     };
   }, [ctx]);
-
-  const handleKeyDown = (event: KeyboardEvent) => {
-    switch (event.key) {
-      case 'ArrowUp':
-        dispatch(movePlayer({ x: 0, y: -6 }));
-        break;
-      case 'ArrowDown':
-        dispatch(movePlayer({ x: 0, y: 6 }));
-        break;
-      case 'ArrowLeft':
-        dispatch(movePlayer({ x: -6, y: 0 }));
-        break;
-      case 'ArrowRight':
-        dispatch(movePlayer({ x: 6, y: 0 }));
-        break;
-      default:
-        break;
-    }
-  };
-
-  useEffect(() => {
-    window.addEventListener('keydown', handleKeyDown);
-
-    return () => {
-      window.removeEventListener('keydown', handleKeyDown);
-    };
-  }, []);
 
   return (
     <article className={'game-page'}>
