+++ packages/client/src/UpdateFunctions/index.ts
@@ -1,0 +1,1 @@
+export { usePlayerControls } from './UsePlayerControls';
